pr review: address all findings from PR #36
Bug fixes:
  - FADT TMR_VAL_EXT was bit 14, should be bit 8 (ACPI spec 6.5 §5.2.9)
  - SPD DDR2 ranks: byte 5 bits[2:0] is 'ranks minus 1' — .max(1) was
    wrong, dual-rank DIMMs decoded as single-rank. Fixed to +1.
  - CPUID max_threads_sharing truncation: 12-bit field (bits 25:14)
    was cast to u8 before division, losing values >255. Use u16.

Safety improvements:
  - SocBootHeader trait methods now 'unsafe fn' — callers must
    acknowledge they're dereferencing raw memory at header_base.
    Updated sunxi impl + all generated call sites with SAFETY comments.
  - CAR teardown: add defensive wbinvd after MTRR re-enable, before
    clearing CD+NW in CR0.
  - phy.rs DQS strobe: add SAFETY comment explaining the intentional
    read from address 0 during receive-enable training.
  - SPD: clear si.dimms[i] when spd_map[i] == 0 to prevent stale
    data on retries.

New crate — fstart-pmio-ich (431 lines):
  Shared ACPI PM I/O abstraction for Intel ICH southbridges:
  - PmIo struct wrapping PMBASE I/O port address
  - TcoIo sub-accessor for TCO register block (PMBASE + 0x60)
  - All PM/GPE/SMI/TCO register offsets and bit definitions from
    coreboot's pmutil.h
  - Named helpers: reset_pm1_status, reset_smi_status,
    reset_gpe0_status, reset_tco_status, global_smi_enable,
    gpe0_mask, alt_gpi_mask, sleep_type, is_s3_resume, poweroff
  - TCO helpers: reset_tco_status (with BOOT_STS ordering), lock
  - Replaces all raw fstart_pio calls + DEFAULT_PMBASE arithmetic
    in the ICH7 driver — net -181 lines from ICH7

API improvements:
  - Network::mac_address returns Result<[u8; 6], ServiceError>
  - X86Config + X86PlatformAcpi: add pmbase field (was hardcoded
    0x500). Default 0x0500 for backward compat. Threaded through
    codegen → fstart-acpi FADT builder.
  - MADT num_cpus default changed from 1 to 0 (runtime-detect
    sentinel).
  - fstart-services: re-export BootLinuxParams and SocBootHeader
    at crate root alongside other service types.

Defensive improvements:
  - ITE8721F: enter_last_byte rejects unknown ports (only 0x2E/0x4E)
  - CK505: init() returns Err(InitFailed) when SmBus not wired
  - SuperIO: init() fails with MissingResource if board requests a
    function that the chip doesn't support (LDN is None)
  - ECAM init: mask off low 20 bits of base address
  - boot_media_values_for_device: return empty Vec instead of panic

Documentation:
  - Pineview: remove stale 'Currently a stub' from raminit doc
  - AArch64: remove stale 3-argument boot_linux docs
  - RTT_NOM: expanded comment (dual-DIMM ODT, not dual-rank)
  - BOOT_PATH_RESET: expanded comment (NB concern, not SB)
  - SMBIOS: clarify 0/Unknown are runtime-detect sentinels
  - MP scatter: document intentional BSP-after-APs ordering
  - hpet_udelay: add TODO for HPET counter-based timing
  - SmBus: remove unnecessary explicit 'return'

Not changed (verified correct after review pushback):
  - Raminit spin-waits: these MCHBAR polls MUST succeed — if the
    hardware doesn't respond, the system is broken and there is no
    recovery path. Timeouts would silently continue with corrupt
    DRAM state, which is worse than hanging.
  - RTT_NOM slot check: dual-DIMM ODT is about bus loading, not
    rank count. Our d0_pop && d1_pop matches coreboot.
  - SIPI trampoline: known TODO, tracked in next steps.

+++ crates/fstart-platform-x86_64/src/cpuid.rs
@@ -153,9 +153,10 @@ pub fn core_thread_count() -> (u8, u8) {
     }
     let r = cpuid(4, 0);
     let max_cores = ((r.eax >> 26) & 0x3F) as u8 + 1;
-    let max_threads_sharing = ((r.eax >> 14) & 0xFFF) as u8 + 1;
+    // Bits 25:14 are 12 bits wide (max 4095) — use u16 to avoid truncation.
+    let max_threads_sharing = ((r.eax >> 14) & 0xFFF) as u16 + 1;
     let threads_per_core = if max_cores > 0 {
-        max_threads_sharing / max_cores
+        (max_threads_sharing / max_cores as u16) as u8
     } else {
         1
     };
